+++ ThirdPartyAdapters/pubmatic/pubmatic/src/main/kotlin/com/google/ads/mediation/pubmatic/PubMaticAdFactory.kt
@@ -16,6 +16,13 @@ import kotlin.time.Duration.Companion.seconds
 interface PubMaticAdFactory {
   fun createPOBInterstitial(context: Context): POBInterstitial
 
+  fun createPOBInterstitial(
+    context: Context,
+    pubId: String,
+    profileId: Int,
+    adUnit: String,
+  ): POBInterstitial
+
   fun createPOBRewardedAd(context: Context): POBRewardedAd
 
   fun createPOBBannerView(context: Context): POBBannerView