+++ ThirdPartyAdapters/pubmatic/pubmatic/src/test/kotlin/com/google/ads/mediation/pubmatic/PubMaticInterstitialAdTests.kt
@@ -84,6 +84,7 @@ class PubMaticInterstitialAdTests {
         mediationInterstitialAdConfiguration,
         mediationAdLoadCallback,
         pubMaticAdFactory,
+        isRtb = true,
       )
       .onSuccess { pubMaticInterstitialAd = it }
   }